backend/requirements.txt
fastapi
uvicorn
stripe
sqlalchemy
requests
python-dotenv